requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.6
discord.py==2.3.2
dnspython==2.4.2
Flask==3.0.0
frozenlist==1.4.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mutf8==1.0.6
mysql-connector-python==8.2.0
protobuf==4.21.12
pycparser==2.22
pymongo==4.6.1
PyNBT==3.1.0
requests==2.31.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
